requirements.txt
flask>=2.3.2
requests>=2.31.0
loguru>=0.7.0
pymongo>=4.3.3
apprise>=1.4.0
pyyaml>=6.0

google-api-python-client~=2.88.0
google-auth~=2.19.1
google_auth_oauthlib~=1.0.0
google-auth-httplib2~=0.1.0
